Demonstrated context binding in various scenarios
I can't find where this topic is touched upon enough to understand the exercise. I spent an hour googling about it but it's not super clear to me. Will call in to a code coach at next session later today but submitted for now, to move on to next task